backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.27.2
yfinance==0.2.50
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.0.0
anthropic==0.40.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0
python-dateutil==2.9.0
greenlet>=3.0.0
